requirements.txt
langchain
langchain-classic
langchain-astradb
langchain-huggingface
langchain-groq
langchain-community
datasets
pypdf
python-dotenv
pandas
flask
prometheus_client